Create P07-Digital_Read_Button.ino
Reads the state of a pushbutton on pin 5. If pressed (HIGH), it turns the LED on pin 13 ON and prints "LED is ON." Otherwise, it turns the LED OFF and prints "LED is OFF".

Using an External Pull-Down Resistor (Preferred)
Left Side → Connect to 5V.
Top Right → Connect to Arduino pin 5.
Bottom Right → Connect to GND through a 10kΩ pull-down resistor.